// src/app/model/ProductInfo.ts
export class ProductInfo{
    productId: string;
    productName: string;
    productPrice: number;
    productStock: number;
    productDescription: string;
    productIcon: string;
    productStatus: number; // 0 :onsale 1:offsale
    categoryType: number;
    createTime: string;
    updateTime: string;

    constructor(){

    }
}